first modularization of hyperoptimizer
